pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vla-paper-digest"
version = "0.1.0"
description = "Daily VLA smoothness paper digest delivered by email via Semantic Scholar, OpenReview and arXiv."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "VLA Paper Digest" }]
license = { text = "MIT" }
dependencies = [
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "pydantic>=2.7.0",
  "jinja2>=3.1.4",
  "tenacity>=8.3.0",
  "python-dotenv>=1.0.1",
  "loguru>=0.7.2",
  "feedparser>=6.0.11",
  "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.2.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
